Disallow phantom type parameters in classes
