--- tests/run/phantom-decls-1.check
@@ -1,4 +1,0 @@
-Boo1
-Boo1.polyfun1
-Boo1
-Boo1.polyfun1
