--- tests/run/phantom-decls-5.check
@@ -1,2 +1,0 @@
-Boo5
-Boo5
